+++ src/main.rs
@@ -12,6 +12,9 @@ use std::str::FromStr;
 
 mod chrono_clock;
 mod cli_helper;
+#[cfg(ical)]
+mod ical_export;
+mod ical_export;
 mod json_storage;
 mod rtw_cli;
 mod rtw_config;